requirements.txt
camelot-py==1.0.0
docx2txt==0.8
langchain-ollama==0.2.2
ollama==0.4.7
pandas==2.2.3
PyMuPDF==1.25.2
PyPDF2==3.0.1
python-dotenv